Replace MasjidActivity with NewMasjidActivity
