+++ JTime-android/src/main/AndroidManifest.xml
@@ -16,10 +16,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activity.MasjidActivity" />
         <activity
-            android:name=".activity.NewMasjidActivity"
-            android:label="@string/title_activity_new_masjid"
+            android:name=".activity.MasjidActivity"
+            android:label="@string/title_activity_masjid"
             android:parentActivityName=".activity.AllMasjidsActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
