+++ JTime-android/src/main/java/com/ismail_s/jtime/android/activity/MasjidActivity.java
@@ -1,42 +1,121 @@
 package com.ismail_s.jtime.android.activity;
 
+
 import android.app.Activity;
+import android.app.Fragment;
+import android.app.FragmentManager;
 import android.os.Bundle;
+import android.support.v13.app.FragmentPagerAdapter;
+import android.support.v13.app.FragmentStatePagerAdapter;
+import android.support.v4.view.ViewPager;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.widget.TextView;
 
-import com.ismail_s.jtime.android.MasjidPojo;
 import com.ismail_s.jtime.android.R;
-import com.ismail_s.jtime.android.RestClient;
 
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.GregorianCalendar;
-import java.util.Locale;
 
 public class MasjidActivity extends Activity {
-    private ArrayList<TextView> dateTextViews = new ArrayList<>();
+    /**
+     * This should be a big number, so that it is as if there are an infinite number of pages.
+     */
+    private static final int NUM_OF_PAGES = 1000;
+
+    /**
+     * The {@link android.support.v4.view.PagerAdapter} that will provide
+     * fragments for each of the sections. We use a
+     * {@link FragmentPagerAdapter} derivative, which will keep every
+     * loaded fragment in memory. If this becomes too memory intensive, it
+     * may be best to switch to a
+     * {@link android.support.v13.app.FragmentStatePagerAdapter}.
+     */
+    private SectionsPagerAdapter mSectionsPagerAdapter;
+
+    /**
+     * The {@link ViewPager} that will host the section contents.
+     */
+    private ViewPager mViewPager;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_masjid);
+        getActionBar().setDisplayHomeAsUpEnabled(true);
+        // Create the adapter that will return a fragment for each of the three
+        // primary sections of the activity.
+        mSectionsPagerAdapter = new SectionsPagerAdapter(getFragmentManager());
+
+        // Set up the ViewPager with the sections adapter.
+        mViewPager = (ViewPager) findViewById(R.id.container);
+        mViewPager.setAdapter(mSectionsPagerAdapter);
+        mViewPager.setCurrentItem(NUM_OF_PAGES / 2);
         String masjidName = getIntent().getStringExtra(Constants.MASJID_NAME);
         TextView masjidNameView = (TextView) findViewById(R.id.masjid_name);
         masjidNameView.setText(masjidName);
-        MasjidPojo masjidTimes = (new RestClient()).getMasjidTimes(masjidName);
-        dateTextViews.add((TextView) findViewById(R.id.fajr_date));
-        dateTextViews.add((TextView) findViewById(R.id.zohar_date));
-        dateTextViews.add((TextView) findViewById(R.id.asr_date));
-        dateTextViews.add((TextView) findViewById(R.id.magrib_date));
-        dateTextViews.add((TextView) findViewById(R.id.esha_date));
-        GregorianCalendar[] times = masjidTimes.getTimes();
-        for (int i = 0; i < dateTextViews.size(); i++) {
-            dateTextViews.get(i).setText(formatCalendar(times[i]));
-        }
+
+
     }
 
-    private String formatCalendar(GregorianCalendar calendar) {
-        SimpleDateFormat formatter = new SimpleDateFormat("HH:mm", Locale.getDefault());
-        return formatter.format(calendar.getTime());
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.menu_masjid, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+
+        //noinspection SimplifiableIfStatement
+        if (id == R.id.action_settings) {
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
+    }
+
+    /**
+     * A {@link FragmentPagerAdapter} that returns a fragment corresponding to
+     * one of the sections/tabs/pages.
+     */
+    public class SectionsPagerAdapter extends FragmentStatePagerAdapter {
+
+        public SectionsPagerAdapter(FragmentManager fm) {
+            super(fm);
+        }
+
+        @Override
+        public Fragment getItem(int position) {
+            // getItem is called to instantiate the fragment for the given page.
+            // Return a NewMasjidFragment (defined as a static inner class below).
+            position = position - (NUM_OF_PAGES / 2);
+            GregorianCalendar date = new GregorianCalendar();
+            date.add(GregorianCalendar.DAY_OF_YEAR, position);
+            return NewMasjidFragment.newInstance(date);
+        }
+
+        @Override
+        public int getCount() {
+            return NUM_OF_PAGES;
+        }
+
+        @Override
+        public CharSequence getPageTitle(int position) {
+            switch (position) {
+                case 0:
+                    return "SECTION 1";
+                case 1:
+                    return "SECTION 2";
+                case 2:
+                    return "SECTION 3";
+            }
+            return null;
+        }
     }
 }
