+++ JTime-android/src/main/res/layout/fragment_masjid.xml
