+++ JTime-android/src/main/res/menu/menu_masjid.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.ismail_s.jtime.android.activity.NewMasjidActivity">
+    tools:context=".activity.MasjidActivity">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
