+++ JTime-android/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="title_activity_new_masjid">NewMasjidActivity</string>
+    <string name="title_activity_masjid">MasjidActivity</string>
     <string name="action_settings">Settings</string>
     <string name="section_format">Hello World from section: %1$s</string>
 </resources>
