+++ JTime-android/src/test/java/com/ismail_s/jtime/android/activity/AllMasjidsActivityTest.java
@@ -52,7 +52,7 @@ public class AllMasjidsActivityTest {
     public void testListActivityStartsMasjidActivityWithCorrectIntent() {
         for (int i = 0; i <= 5; i++) {
             clickOnItem(i);
-            Intent expectedIntent = new Intent(activity, NewMasjidActivity.class);
+            Intent expectedIntent = new Intent(activity, MasjidActivity.class);
             expectedIntent.putExtra(Constants.MASJID_NAME, masjids[i]);
             Intent actualIntent = Shadows.shadowOf(activity).getNextStartedActivity();
             assertEquals(expectedIntent, actualIntent);
